backend/requirements.txt
# StudyMate Backend - Local Development Dependencies
# All services consolidated for local development

# FastAPI and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.5.0

# Database and Storage
asyncpg==0.29.0
supabase>=2.19.0
motor==3.3.2
pymongo==4.6.0

# AI and Processing
groq==0.32.0
google-generativeai==0.3.2

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client and Utilities
httpx>=0.25.2
python-dotenv==1.0.0
redis==5.0.1
